+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/model/BodyWithContentType.java
@@ -6,13 +6,13 @@ package de.gematik.test.tiger.mockserver.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.nio.charset.Charset;
-import java.util.Objects;
+import lombok.EqualsAndHashCode;
 
 /*
  * @author jamesdbloom
  */
+@EqualsAndHashCode
 public abstract class BodyWithContentType<T> extends Body<T> {
-  private int hashCode;
   protected final MediaType contentType;
 
   public BodyWithContentType(Type type, MediaType contentType) {
@@ -40,31 +40,5 @@ public abstract class BodyWithContentType<T> extends Body<T> {
   @Override
   public String getContentType() {
     return (contentType != null ? contentType.toString() : null);
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    if (hashCode() != o.hashCode()) {
-      return false;
-    }
-    if (!super.equals(o)) {
-      return false;
-    }
-    BodyWithContentType<?> that = (BodyWithContentType<?>) o;
-    return Objects.equals(contentType, that.contentType);
-  }
-
-  @Override
-  public int hashCode() {
-    if (hashCode == 0) {
-      hashCode = Objects.hash(super.hashCode(), contentType);
-    }
-    return hashCode;
   }
 }