+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/model/HttpRequest.java
@@ -307,7 +307,7 @@ public class HttpRequest extends RequestDefinition implements HttpMessage<HttpRe
   }
 
   public HttpRequest shallowClone() {
-    return not(request(), not)
+    return request()
         .setMethod(method)
         .setPath(path)
         .setPathParameters(pathParameters)
@@ -327,7 +327,7 @@ public class HttpRequest extends RequestDefinition implements HttpMessage<HttpRe
 
   @SuppressWarnings("MethodDoesntCallSuperMethod")
   public HttpRequest clone() {
-    return not(request(), not)
+    return request()
         .setMethod(method)
         .setPath(path)
         .setPathParameters(pathParameters != null ? pathParameters.clone() : null)
@@ -343,10 +343,7 @@ public class HttpRequest extends RequestDefinition implements HttpMessage<HttpRe
         .setClientCertificateChain(
             clientCertificateChain != null && !clientCertificateChain.isEmpty()
                 ? clientCertificateChain.stream()
-                    .map(
-                        c ->
-                            MockserverX509CertificateWrapper.with(
-                                (java.security.cert.X509Certificate) c.certificate()))
+                    .map(c -> MockserverX509CertificateWrapper.with(c.certificate()))
                     .toList()
                 : null)
         .setSocketAddress(socketAddress)