+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/model/HttpResponse.java
@@ -283,14 +283,6 @@ public class HttpResponse extends Action<HttpResponse>
     }
   }
 
-  private Cookies getOrCreateCookies() {
-    if (this.cookies == null) {
-      this.cookies = new Cookies();
-      this.hashCode = 0;
-    }
-    return this.cookies;
-  }
-
   public HttpResponse withCookies(Cookies cookies) {
     if (cookies == null || cookies.isEmpty()) {
       this.cookies = null;
@@ -329,7 +321,6 @@ public class HttpResponse extends Action<HttpResponse>
         .withBody(body)
         .withHeaders(headers != null ? headers.clone() : null)
         .withCookies(cookies != null ? cookies.clone() : null)
-        .withDelay(getDelay())
         .withStreamId(streamId);
   }