+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/model/RequestDefinition.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 /*
  * @author jamesdbloom
  */
-public abstract class RequestDefinition extends Not {
+public abstract class RequestDefinition extends ObjectWithJsonToString {
 
   private String logCorrelationId;
 
@@ -24,10 +24,6 @@ public abstract class RequestDefinition extends Not {
   }
 
   public abstract RequestDefinition shallowClone();
-
-  public RequestDefinition cloneWithLogCorrelationId() {
-    return shallowClone().withLogCorrelationId(getLogCorrelationId());
-  }
 
   @Override
   public boolean equals(Object o) {