+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TigerProxyExamplesTest.java
@@ -5,6 +5,7 @@
 package de.gematik.test.tiger.proxy;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static de.gematik.rbellogger.data.RbelElementAssertion.assertThat;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
@@ -83,8 +84,9 @@ class TigerProxyExamplesTest {
           .asString();
       TigerProxyTestHelper.waitUntilMessageListInProxyContainsCountMessages(tigerProxy, 2);
 
-      assertThat(tigerProxy.getRbelMessagesList().get(1).getRawStringContent())
-          .contains("barschmoolildu");
+      assertThat(tigerProxy.getRbelMessagesList().get(1))
+        .extractChildWithPath("$.body")
+        .hasStringContentEqualTo("barschmoolildu");
     }
   }